+++ lib/active_interaction/base.rb
@@ -121,20 +121,14 @@ module ActiveInteraction
       #
       # @return (see .filters)
       #
-      # @raise [ArgumentError] If both `:only` and `:except` are given.
-      #
       # @!visibility public
       def import_filters(klass, options = {})
-        if options.key?(:only) && options.key?(:except)
-          fail ArgumentError, 'given both :only and :except'
-        end
-
         only = options[:only]
         except = options[:except]
 
         other_filters = klass.filters.dup
-        other_filters.select! { |k, _| only.include?(k) } if only
-        other_filters.reject! { |k, _| except.include?(k) } if except
+        other_filters.select! { |k, _| [*only].include?(k) } if only
+        other_filters.reject! { |k, _| [*except].include?(k) } if except
 
         filters.merge!(other_filters)
       end